services/users/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
Authlib==1.5.0
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.1
identify==2.6.9
idna==3.10
itsdangerous==2.2.0
jiter==0.9.0
jose==1.0.0
jsonpath-python==1.0.6
kombu==5.5.0
mistralai==1.5.1
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic>=2.0
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
